add one more testcase to AEZ
Which captured subtle bug found in aezeed.
